Let hypermedia be a proper CommonJS module
Just as angular modules (angular-messages, ui-router) published on NPM,
this module should be a proper commonjs module. This way it can be
`require()`d just as those modules.

Just as the official angular modules, this creates an index.js that does the
actual export, guarded by a commen check for a commonjs environment.